updated enviorment file
