+++ README.md
@@ -1,2 +1,4 @@
 # MiniProject
 MiniProject for Full stack Development 
+
+run enviorment.yml with command `conda env create -f enviorment.yml` to create the enviorment
