+++ enviroment.yml
@@ -1,0 +1,23 @@
+name: cse108_mini_project
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - python=3.9  # Specify your Python version here
+  - alembic=1.13.1
+  - flask=2.2.5
+  - flask-migrate=4.0.7
+  - flask-sqlalchemy=3.0.3
+  - flask-wtf=1.0.1
+  - flask-cors=3.0.10
+  - wtforms=2.3.3
+  - jinja2=3.1.3
+  - sqlalchemy=2.0.29
+  - werkzeug=3.0.2
+  - bcrypt=4.1.2
+  - flask-bcrypt=1.0.1
+  - flask-login=0.6.3
+  - requests=2.31.0
+  - flask-admin=1.6.1
+  # - pip:
+  #   - flask-bcrypt==1.0.1  # Use pip if some packages are not available in Conda
